fix: Show remaining countdown if less than a minute
