--- src/components/common/Countdown/index.test.ts
@@ -1,23 +1,0 @@
-import { _getCountdown } from '.'
-
-describe('getCountdown', () => {
-  it('should convert 0 seconds to 0 days, 0 hours, and 0 minutes', () => {
-    const result = _getCountdown(0)
-    expect(result).toEqual({ days: 0, hours: 0, minutes: 0 })
-  })
-
-  it('should convert 3600 seconds to 0 days, 1 hour, and 0 minutes', () => {
-    const result = _getCountdown(3600)
-    expect(result).toEqual({ days: 0, hours: 1, minutes: 0 })
-  })
-
-  it('should convert 86400 seconds to 1 day, 0 hours, and 0 minutes', () => {
-    const result = _getCountdown(86400)
-    expect(result).toEqual({ days: 1, hours: 0, minutes: 0 })
-  })
-
-  it('should convert 123456 seconds to 1 day, 10 hours, and 17 minutes', () => {
-    const result = _getCountdown(123456)
-    expect(result).toEqual({ days: 1, hours: 10, minutes: 17 })
-  })
-})
